requirements.txt
loguru
openpyxl
requests
pandas
PyExecJS
beautifulsoup4
